# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Perf Demo API","description":"APM 성능 병목 분석 교육용 데모 API\n\n모든 API에 ?optimized=true/false 파라미터로 Before/After 전환 가능\n\n| Scenario | API | 병목 | 수정 |\n|----------|-----|------|------|\n| 1 | GET /api/orders/{id} | Thread.sleep(3s) | sleep 제거 |\n| 2 | GET /api/orders/search | LIKE '%keyword%' | 인덱스 + prefix LIKE |\n| 3 | GET /api/orders/detail-list | N+1 쿼리 | JOIN 한 방 |\n| 4 | POST /api/orders/process | 동기 로깅 | DEBUG + Async |\n| 5 | GET /api/orders/report | 앱 루프 | DB 집계 |","version":"1.0.0"},"servers":[{"url":"http://localhost:18081","description":"Generated server url"}],"tags":[{"name":"주문 API","description":"성능 병목 분석 교육용 API (?optimized=true/false로 Before/After 전환)"}],"paths":{"/api/orders/process":{"post":{"tags":["주문 API"],"summary":"Scenario 4: 로깅 병목","description":"Before: 동기 대용량 로깅 / After: DEBUG 레벨 + AsyncAppender","operationId":"processOrders","parameters":[{"name":"count","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1000}},{"name":"optimized","in":"query","description":"true=최적화, false=병목","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}},"/api/orders/{orderId}":{"get":{"tags":["주문 API"],"summary":"Scenario 1: 단건 조회","description":"Before: Thread.sleep(3000) 병목 / After: sleep 제거","operationId":"getOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"string"}},{"name":"optimized","in":"query","description":"true=최적화, false=병목","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseOrderResponseDTO"}}}}}}},"/api/orders/search":{"get":{"tags":["주문 API"],"summary":"Scenario 2: 키워드 검색","description":"Before: LIKE '%keyword%' 풀스캔 / After: 인덱스 + prefix LIKE","operationId":"searchOrders","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}},{"name":"optimized","in":"query","description":"true=최적화, false=병목","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListOrderResponseDTO"}}}}}}},"/api/orders/report":{"get":{"tags":["주문 API"],"summary":"Scenario 5: 쿼리 간 서비스 병목","description":"Before: Query→루프→Query / After: DB 집계 쿼리","operationId":"getOrderReport","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"optimized","in":"query","description":"true=최적화, false=병목","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListOrderReportResponseDTO"}}}}}}},"/api/orders/detail-list":{"get":{"tags":["주문 API"],"summary":"Scenario 3: N+1 쿼리","description":"Before: 1+N 건별 조회 / After: JOIN 한 방 쿼리","operationId":"getOrdersWithDetails","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":50}},{"name":"optimized","in":"query","description":"true=최적화, false=병목","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListOrderWithDetailsResponseDTO"}}}}}}}},"components":{"schemas":{"ApiResponseString":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"string"},"error":{"type":"string"}}},"ApiResponseOrderResponseDTO":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/OrderResponseDTO"},"error":{"type":"string"}}},"OrderResponseDTO":{"type":"object","properties":{"orderId":{"type":"string","description":"주문 ID","example":"ORD-00001"},"customerId":{"type":"string","description":"고객 ID"},"customerName":{"type":"string","description":"고객명"},"orderStatus":{"type":"string","description":"주문 상태"},"totalAmount":{"type":"integer","description":"총 금액","format":"int64"},"orderDate":{"type":"string","description":"주문일"}},"description":"주문 응답"},"ApiResponseListOrderResponseDTO":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/OrderResponseDTO"}},"error":{"type":"string"}}},"ApiResponseListOrderReportResponseDTO":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/OrderReportResponseDTO"}},"error":{"type":"string"}}},"OrderReportResponseDTO":{"type":"object","properties":{"orderStatus":{"type":"string","description":"주문 상태"},"orderCount":{"type":"integer","description":"건수","format":"int32"},"totalAmountSum":{"type":"integer","description":"총 금액 합계","format":"int64"},"productNames":{"type":"string","description":"상품 목록 (LISTAGG)"}},"description":"주문 리포트 응답 (Mixed 병목 시나리오)"},"ApiResponseListOrderWithDetailsResponseDTO":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/OrderWithDetailsResponseDTO"}},"error":{"type":"string"}}},"OrderDetailResponseDTO":{"type":"object","properties":{"detailId":{"type":"string","description":"상세 ID"},"orderId":{"type":"string","description":"주문 ID"},"productId":{"type":"string","description":"상품 ID"},"productName":{"type":"string","description":"상품명"},"quantity":{"type":"integer","description":"수량","format":"int32"},"unitPrice":{"type":"integer","description":"단가","format":"int64"},"lineAmount":{"type":"integer","description":"소계","format":"int64"}},"description":"주문 상세 응답"},"OrderWithDetailsResponseDTO":{"type":"object","properties":{"orderId":{"type":"string","description":"주문 ID"},"customerName":{"type":"string","description":"고객명"},"orderStatus":{"type":"string","description":"주문 상태"},"totalAmount":{"type":"integer","description":"총 금액","format":"int64"},"details":{"type":"array","description":"주문 상세 목록","items":{"$ref":"#/components/schemas/OrderDetailResponseDTO"}}},"description":"주문 + 상세 목록 응답 (N+1 시나리오)"}}}}